backend/requirements.txt
fastapi[standard]

# for mri brain extraction
opencv-python-headless
antspyx==0.5.4
antspynet
tf-keras # required for antspynet